splitted guards to auth_manager
