+++ src/iot/mod.rs
@@ -1,9 +1,9 @@
 ///All logic which is not web-based.
-pub mod door_control;
+mod door_control;
 pub use door_control::DoorControl;
 
-pub mod bell_button;
-pub use bell_button::BellButton;
+mod bell_button;
+use bell_button::BellButton;
 
-pub mod event_handler;
+mod event_handler;
 pub use event_handler::event_loop;
