+++ src/requests/user_auth.rs
@@ -1,7 +1,7 @@
 use super::index_view::*;
 use crate::db_entry::DbConn;
-use crate::guards::GuardManager;
 use crate::template_contexts::{LoginContext, Message};
+use crate::utils::auth_manager::AuthManager;
 use rocket::http::Cookies;
 use rocket::request::{FlashMessage, Form};
 use rocket::response::{Flash, Redirect};
@@ -30,7 +30,7 @@ pub fn post_login_data(
     conn: DbConn,
     cookies: Cookies,
 ) -> Result<Redirect, Flash<Redirect>> {
-    let user = match GuardManager::auth(&conn, cookies, &user_data.name, &user_data.pw) {
+    let user = match AuthManager::auth(&conn, cookies, &user_data.name, &user_data.pw) {
         Err(e) => return Err(Flash::error(Redirect::to(uri!(get_login)), e.to_string())),
         Ok(user) => user,
     };
@@ -46,7 +46,7 @@ pub fn post_login_data(
 /// Get logout to destroy the user-cookie
 #[get("/logout")]
 pub fn get_logout(cookies: Cookies) -> Flash<Redirect> {
-    GuardManager::destroy_user_cookie(cookies);
+    AuthManager::destroy_user_cookie(cookies);
     return Flash::success(
         Redirect::to(uri!(get_login)),
         "Sie wurden erfolgreich ausgeloggt",
