+++ src/utils/auth_manager.rs
@@ -1,11 +1,9 @@
 /// Are used for the authentification and authorization.
-use crate::crypto;
+use super::crypto;
 use crate::db_entry::{rusqlite, DbConn, HashEntry, UserEntry};
 use blake2::{Blake2b, Digest};
 use passwords::{analyzer, scorer};
-use rocket::http::{Cookie, Cookies, Status};
-use rocket::request::{self, FromRequest, Request};
-use rocket::Outcome;
+use rocket::http::{Cookie, Cookies};
 use std::fmt;
 
 /// All errors which could happen during user creation and authentification.
@@ -45,9 +43,9 @@ impl fmt::Display for AuthError {
 }
 
 /// Used for user creation and authentification
-pub struct GuardManager {}
+pub struct AuthManager {}
 
-impl GuardManager {
+impl AuthManager {
     /// Checks whether the password is secure or errors if it is weak
     pub fn check_password(pw: &str) -> Result<(), AuthError> {
         if scorer::score(&analyzer::analyze(pw)) < 80f64 {
@@ -109,7 +107,7 @@ impl GuardManager {
             return Err(AuthError::WrongPassword);
         }
 
-        GuardManager::write_user_cookie(&user, cookies)?;
+        AuthManager::write_user_cookie(&user, cookies)?;
 
         Ok(user)
     }
@@ -130,81 +128,5 @@ impl GuardManager {
     /// Destroys the private encrypted user cookie.
     pub fn destroy_user_cookie(mut cookies: Cookies) {
         cookies.remove_private(Cookie::named("user"));
-    }
-}
-
-/// A guard which allows all authentificated users.
-pub struct UserGuard {
-    pub user: UserEntry,
-}
-
-impl UserGuard {
-    pub fn is_user(&self) -> bool {
-        self.user.user_type.is_user()
-    }
-
-    pub fn is_admin(&self) -> bool {
-        self.user.user_type.is_admin()
-    }
-}
-
-impl<'a, 'r> FromRequest<'a, 'r> for UserGuard {
-    type Error = AuthError;
-
-    /// Checks for valid user-cookie in a request
-    fn from_request(request: &'a Request<'r>) -> request::Outcome<UserGuard, AuthError> {
-        return request
-            .cookies()
-            .get_private("user")
-            .map_or(Outcome::Forward(()), |cookie| {
-                match serde_json::from_str(cookie.value()) {
-                    Ok(user) => Outcome::Success(UserGuard { user }),
-                    Err(e) => Outcome::Failure((Status::BadRequest, AuthError::from(e))),
-                }
-            });
-    }
-}
-
-/// A guard which allows only users.
-pub struct OnlyUserGuard {
-    pub user: UserEntry,
-}
-
-impl<'a, 'r> FromRequest<'a, 'r> for OnlyUserGuard {
-    type Error = AuthError;
-
-    /// Checks if a valid client is a user
-    fn from_request(request: &'a Request<'r>) -> request::Outcome<OnlyUserGuard, AuthError> {
-        let user_guard = request.guard::<UserGuard>()?;
-
-        if user_guard.is_user() {
-            Outcome::Success(OnlyUserGuard {
-                user: user_guard.user,
-            })
-        } else {
-            Outcome::Forward(())
-        }
-    }
-}
-
-/// A guard which allows only administrators.
-pub struct AdminGuard {
-    pub user: UserEntry,
-}
-
-impl<'a, 'r> FromRequest<'a, 'r> for AdminGuard {
-    type Error = AuthError;
-
-    /// Checks if a valid client is an admin
-    fn from_request(request: &'a Request<'r>) -> request::Outcome<AdminGuard, AuthError> {
-        let user_guard = request.guard::<UserGuard>()?;
-
-        if user_guard.is_admin() {
-            Outcome::Success(AdminGuard {
-                user: user_guard.user,
-            })
-        } else {
-            Outcome::Forward(())
-        }
     }
 }
